refactored using redux-toolkit, moved to redux dir
