--- part6/redux-anecdotes/src/reducers/anecdoteReducer.js
@@ -1,54 +1,0 @@
-const anecdotesAtStart = [
-  "If it hurts, do it more often",
-  "Adding manpower to a late software project makes it later!",
-  "The first 90 percent of the code accounts for the first 90 percent of the development time...The remaining 10 percent of the code accounts for the other 90 percent of the development time.",
-  "Any fool can write code that a computer can understand. Good programmers write code that humans can understand.",
-  "Premature optimization is the root of all evil.",
-  "Debugging is twice as hard as writing the code in the first place. Therefore, if you write the code as cleverly as possible, you are, by definition, not smart enough to debug it.",
-];
-
-const getId = () => (100000 * Math.random()).toFixed(0);
-
-const asObject = (anecdote) => {
-  return {
-    content: anecdote,
-    id: getId(),
-    votes: 0,
-  };
-};
-
-const initialState = anecdotesAtStart.map(asObject);
-
-const reducer = (state = initialState, action) => {
-  const payload = action.payload;
-  switch (action.type) {
-    case "UPVOTE":
-      return state.map((anecdote) =>
-        anecdote.id != payload.id
-          ? anecdote
-          : { ...anecdote, votes: anecdote.votes + 1 },
-      );
-    case "ADD":
-      return [...state, payload];
-    default:
-      return state;
-  }
-};
-
-const upvote = (id) => {
-  return {
-    type: "UPVOTE",
-    payload: { id },
-  };
-};
-
-const createAnecdote = (content) => {
-  return {
-    type: "ADD",
-    payload: asObject(content),
-  };
-};
-
-export default reducer;
-
-export { upvote, createAnecdote };
